refactor: include insert controller in separate files
